Continued refactoring and correction to the array order for stones in item form (oops)
