+++ TFC_Shared/src/TFC/API/Constant/Global.java
@@ -1,10 +1,13 @@
-package TFC.API;
+package TFC.API.Constant;
 
-public class BlockTypes
+import com.google.common.collect.ObjectArrays;
+
+public interface Global
 {
 	/* Stone Types */
 	public static final String[] STONE_IGEX = {"Rhyolite", "Basalt", "Andesite", "Dacite"};
 	public static final String[] STONE_IGIN = {"Granite", "Diorite", "Gabbro"};
 	public static final String[] STONE_MM   = {"Quartzite", "Slate", "Phyllite", "Schist", "Gneiss", "Marble"};
 	public static final String[] STONE_SED  = {"Siltstone", "Mudstone", "Shale", "Claystone", "Rock Salt", "Limestone", "Conglomerate", "Dolomite", "Chert", "Chalk"};
+	public static final String[] STONE_ALL  = ObjectArrays.concat(ObjectArrays.concat(STONE_IGIN, STONE_SED, String.class), ObjectArrays.concat(STONE_IGEX, STONE_MM, String.class), String.class);
 }
