+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockIgEx.java
@@ -15,7 +15,7 @@ import net.minecraft.world.Explosion;
 import net.minecraft.world.World;
 import TFC.TFCBlocks;
 import TFC.TFCItems;
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import TFC.Core.Helper;
 import TFC.Core.TFC_Core;
 import TFC.Items.Tools.ItemChisel;
@@ -29,7 +29,7 @@ public class BlockIgEx extends BlockStone
 		super(i, material, id);
 		this.dropBlock = TFCBlocks.StoneIgExCobble.blockID;
 
-        names = BlockTypes.STONE_IGEX;
+        names = Global.STONE_IGEX;
         icons = new Icon[names.length];
 	}
