+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockIgInSmooth.java
@@ -3,7 +3,7 @@ package TFC.Blocks.Terrain;
 import java.util.List;
 import java.util.Random;
 import TFC.TerraFirmaCraft;
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import TFC.Blocks.BlockTerra;
 import TFC.Core.Helper;
 import TFC.Core.TFC_Core;
@@ -54,7 +54,7 @@ public class BlockIgInSmooth extends BlockSmooth
 		super(i, Material.rock);
 		this.setCreativeTab(CreativeTabs.tabBlock);
 
-        names = BlockTypes.STONE_IGIN;
+        names = Global.STONE_IGIN;
 		icons = new Icon[names.length];
 	}
 }
