+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockMM.java
@@ -15,7 +15,7 @@ import net.minecraft.world.Explosion;
 import net.minecraft.world.World;
 import TFC.TFCBlocks;
 import TFC.TFCItems;
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import TFC.Core.TFC_Core;
 import TFC.Items.Tools.ItemChisel;
 import TFC.Items.Tools.ItemHammer;
@@ -30,7 +30,7 @@ public class BlockMM extends BlockStone
         super(i, material, id);
         this.dropBlock = TFCBlocks.StoneMMCobble.blockID;
 
-        names = BlockTypes.STONE_MM;
+        names = Global.STONE_MM;
         icons = new Icon[names.length];
     }
