+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockMMCobble.java
@@ -13,7 +13,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.Icon;
 import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.world.World;
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import TFC.Blocks.BlockTerra;
 import TFC.Core.Helper;
 import TFC.Core.TFC_Sounds;
@@ -31,7 +31,7 @@ public class BlockMMCobble extends BlockCobble
 
     	fallInstantly = false;
 
-        names = BlockTypes.STONE_MM;
+        names = Global.STONE_MM;
     	icons = new Icon[names.length];
     }
