+++ TFC_Shared/src/TFC/Items/ItemBlocks/ItemIgEx.java
@@ -1,6 +1,6 @@
 package TFC.Items.ItemBlocks;
 
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import net.minecraft.client.renderer.texture.IconRegister;
 
 public class ItemIgEx extends ItemTerraBlock
@@ -8,7 +8,7 @@ public class ItemIgEx extends ItemTerraBlock
 	public ItemIgEx(int i) 
 	{
 		super(i);
-		MetaNames = BlockTypes.STONE_IGEX;
+		MetaNames = Global.STONE_IGEX;
 	}
 	
 	@Override
