+++ TFC_Shared/src/TFC/Items/ItemBlocks/ItemIgIn.java
@@ -1,6 +1,6 @@
 package TFC.Items.ItemBlocks;
 
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import net.minecraft.client.renderer.texture.IconRegister;
 
 public class ItemIgIn extends ItemTerraBlock
@@ -8,7 +8,7 @@ public class ItemIgIn extends ItemTerraBlock
 	public ItemIgIn(int i) 
 	{
 		super(i);
-		MetaNames = BlockTypes.STONE_IGIN;
+		MetaNames = Global.STONE_IGIN;
 	}
 	
 	@Override
