+++ TFC_Shared/src/TFC/Items/ItemBlocks/ItemMM.java
@@ -1,6 +1,6 @@
 package TFC.Items.ItemBlocks;
 
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import net.minecraft.client.renderer.texture.IconRegister;
 
 public class ItemMM extends ItemTerraBlock
@@ -8,7 +8,7 @@ public class ItemMM extends ItemTerraBlock
 	public ItemMM(int i) 
 	{
 		super(i);
-		MetaNames = BlockTypes.STONE_MM;
+		MetaNames = Global.STONE_MM;
 	}
 	
 	@Override
