+++ TFC_Shared/src/TFC/Items/ItemBlocks/ItemSed.java
@@ -1,6 +1,6 @@
 package TFC.Items.ItemBlocks;
 
-import TFC.API.BlockTypes;
+import TFC.API.Constant.Global;
 import net.minecraft.client.renderer.texture.IconRegister;
 
 public class ItemSed extends ItemTerraBlock
@@ -8,7 +8,7 @@ public class ItemSed extends ItemTerraBlock
 	public ItemSed(int i) 
 	{
 		super(i);
-		MetaNames = BlockTypes.STONE_SED;
+		MetaNames = Global.STONE_SED;
 	}
 	@Override
 	public void registerIcons(IconRegister registerer)
